Change how compiler features are rolled out
This change introduces a new mechnism to roll out compiler
features that are intended to be made the eventual default
behavior. Instead of introducing a flag that can be confused
with configuration option will always be options, all such
features are now enabled or disabled through "featureFlag".

Relnote: """Changed how features being rolled out are
enabled and disabled. Features, such as strong skipping and
non-skipping group optimizations are now enabled through the
"featureFlag" option instead of their own option.

A feature with a feature flag start off initially disabled
and will eventually become enabled by default.

For example, to enable strong skipping use:

  -P plugin:androidx.compose.compiler.plugins.kotlin:featureFlag=StrongSkipping

To enable the non-skipping groups optimization use:

  -P plugin:androidx.compose.compiler.plugins.kotlin:featureFlag=OptimizeNonSkippingGroups

the options strongSkipping, experimentialStrongSkipping,
and nonSkippingGroupOptimization are now all depricated
and will report an warning if used but are still honored.

A feature flag can be removed if its name is prefixed
with a "-". For example:

  -P plugin:androidx.compose.compiler.plugins.kotlin:featureFlag=-StrongSkipping

will disable strong skipping.

Using a feature that is unknown to the plugin will
report a warning.

If a feature is enabled that is enabled by default or
disabled and is disabled by default, a warning will be
issued.
"""

Test: ./gradlew :compose:c:c-h:i-t:check
Change-Id: Iab20b3f0e5e818a98f7952bbe1e1ebbd2b8251c7 ( https://android-review.googlesource.com/q/Iab20b3f0e5e818a98f7952bbe1e1ebbd2b8251c7 )

Moved from: https://github.com/androidx/androidx/commit/881d7830d2ee1018c2b7be923a17fcae1f834c55