[Analysis API] Provide PSI-based implementation of containing symbols
The compiler implementation `getContainingClassSymbol()` always looks up
the containing class by its name. This breaks when there are multiple
classes in the same package. This is a rare occasion during compilation,
but it can often happen in the IDE during refactorings.

After the change, `getContainingClassSymbol()` delegates to the
'FirProvider', where its LL implementation firstly tries to find the
parent class using the PSI and only delegates to the (default) compiler
implementation if it fails to find one.

^KT-61516 Fixed
^KT-67963 Fixed