[FIR] Recreate LocalScope for primary constructor value parameters in
...BodyResolveContext to fix a problem with incorrect symbols
collected during completion (KT-71174).

What happened in details:

1. Context is collected using `KaScopeProvider#scopeContext`
2. First, while walking through the file, value parameter symbols
with "source" module are added into the local scope.
3. Then, `ContextCollectorVisitor#visitConstructor` gets to the point
when it calls `context.forConstructorParameters` and iterates over
parameters.
4. Once it finds an element matching the current position or gets
`FilterResponse.STOP`, it makes a snapshot of collected context
(see `computeContext()` method). In our case, it happens when on 2d
value parameter.
5. But here's the thing: `ContextCollectorVisitor` walks through
dangling file, not original source one, and because of that value
parameters in local scope, which previously were source, are rewritten
with parameters from dangling file.
But we won't see a change in the 3rd parameter, because, as already
said, snapshot had been made when visiting the 2nd parameter.
6. `FirCallableCompletionContributor#isUninitializedCallable` is
responsible for filtering out uninitialized symbols from completion
from the IDE side. It relies on modules equality, and because of that
fails to filter out 3rd symbol.
7. Why does it work in case of functions and secondary constructors?
Because local scope there is recreated before iterating over
parameters (see `BodyResolveContext.forFunctionBody`).
And it's not clear to me why primary constructor doesn't do the same
and instead takes existing scope.
8. So, that's the fix: bring everything in accordance.

^KT-71174