[Java interop] Fix nullability qualifier computation with overrides
In the given test, the parameter of `ConcurrentHashMap.put`
is annotated as not-null using an external annotation, meaning
its qualifier for enhancement is
{nullability=NOT_NULL, definitelyNotNull=true,
isNullabilityQualifierForWarning=true}.
To compute the final qualifier for the parameter of
`ConcurrentHashMap.put`, we consider its own qualifier and the
qualifiers of its overrides.
ConcurrentHashMap indirectly @PurelyImplements MutableMap, so
among `put`'s overrides is `MutableMap.put` whose parameter is
non-flexible `K`.
From this we derive the qualifier {nullability=NOT_NULL,
definitelyNotNull=false, isNullabilityQualifierForWarning=false}.
Before this commit, the algorithm for consolidating the qualifiers
looked at the `nullability` and `definitelyNotNull` values in isolation
and produced {nullability=NOT_NULL, definitelyNotNull=true,
isNullabilityQualifierForWarning=false} because there was _some_
enhancement to NOT_NULL with error severity and _some_ qualifier was
`definitelyNotNull`.
In the given example, this led to an error-severity type mismatch error.
In this commit, the algorithm is reworked to distinguish NOT_NULL
and DEFINITELY_NOT_NULL and compute the finer-grained nullability in one
step.
The nullability is computed for error and warning severity, and in the
case of stateForWarnings == DEFINITELY_NOT_NULL &&
stateForErrors == NOT_NULL, we return DEFINITELY_NOT_NULL with warning
severity.

#KT-67560 Fixed