Fix false-positive TYPE_MISMATCH in nested lambda inside PCLA
See unitElvisLikeWithRun.kt

In K2, we've got the following CS state before starting `elvisLike`
completion:
- Ev – doesn't have proper constraints yet
- For Tv, we've got a lower constraint Box<Ev> <: Tv
(thus TV is related to an outer Ev variable)
- For Kv, we've got String <: Kv and Rv <: Kv
- For Rv, we only have Rv <: Kv

Thus, we considered Tv as related to outer variables,
 so we were not able to fix it and start straightforward lambda
 analysis for `myRun { ... }`.

On the other hand, we didn't consider Kv and Rv as related
to outer type variables, so we would fix them eagerly to `String`.
And then we forcefully start resolving the lambda with semi-fixed
Tv := Box<Ev>, but Unit may not be coerced to String.

The idea of the fix is that we should treat Tv and Rv as related
because Rv should not be fixed before Tv,
as we do in many other cases like related-to-top-level type
(see TypeVariableDependencyInformationProvider.computeRelatedToTopLevelType).

Thus, after that we do that, all the Ev, Tv, Kv, Rv left not fixed
at the point of the lambda analysis, so we just analyze the lambda
with semifixed Tv := Box<Ev> and get new constraints:
- Unit <: Rv
- Unit <: Kv

So, once we start fixing all the type variables (after PCLA),
we would fix Kv to Any (CST(Unit, String)) as it's done in K1.

NB: Basically, we assume that we might postpone fixation for all
type variables of nested calls inside PCLA, but it's likely would lead
to performance problems.

^KT-72238 Fixed